Services/EncryptionService/requirements.txt
cryptography
argparse
grpcio_tools